[SofaExporter] Move bindings to its owner
