+++ applications/plugins/SofaPython/Binding.cpp
@@ -58,8 +58,6 @@
 #include "Binding_SparseGridTopology.h"
 #include "Binding_SubsetMultiMapping.h"
 #include "Binding_VisualModel.h"
-#include "Binding_OBJExporter.h"
-#include "Binding_STLExporter.h"
 #include "Binding_DataEngine.h"
 #include "PythonFactory.h"
 
@@ -121,8 +119,6 @@ void bindSofaPythonModule(PyObject * module)
     SP_ADD_CLASS_IN_FACTORY(GridTopology,sofa::component::topology::GridTopology)
     SP_ADD_CLASS_IN_FACTORY(SparseGridTopology,sofa::component::topology::SparseGridTopology)
     SP_ADD_CLASS_IN_FACTORY(RegularGridTopology,sofa::component::topology::RegularGridTopology)
-    SP_ADD_CLASS_IN_FACTORY(OBJExporter,sofa::component::misc::OBJExporter)
-    SP_ADD_CLASS_IN_FACTORY(STLExporter,sofa::component::misc::STLExporter)
     SP_ADD_CLASS_IN_FACTORY(PythonScriptController,sofa::component::controller::PythonScriptController)
     SP_ADD_CLASS_IN_FACTORY(PythonScriptDataEngine,sofa::component::controller::PythonScriptDataEngine)
     SP_ADD_CLASS_IN_FACTORY(PointSetTopologyModifier,sofa::component::topology::PointSetTopologyModifier)
